LayoutTests/imported/w3c/web-platform-tests/css/css-conditional/container-queries/container-name-tree-scoped-expected.txt


PASS Outer scope query should not match container-name set by :host rule in shadow tree
PASS Outer scope query should not match container-name set by ::slotted rule in shadow tree
FAIL Inner scope query should match container-name set by :host rule in shadow tree assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
FAIL Inner scope query containing ::slotted should match container-name set by :host rule in shadow tree assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
